fix: types and minor issues
