+++ apps/api/src/auth/strategy/google.strategy.ts
@@ -5,6 +5,7 @@ import { ConfigService } from '@nestjs/config';
 import { UsersService } from '../../users/users.service';
 import { UserServicesService } from '../../user-services/user-services.service';
 import { Request } from 'express';
+import { parsePlatformFromState } from 'src/utils/parsePlatform';
 
 @Injectable()
 export class GoogleStrategy extends PassportStrategy(Strategy, 'google') {
@@ -28,16 +29,7 @@ export class GoogleStrategy extends PassportStrategy(Strategy, 'google') {
     refreshToken: string,
     profile: Profile,
   ) {
-    let platform = 'web';
-    try {
-      const state = req.query.state;
-      if (typeof state === 'string') {
-        const parsed = JSON.parse(state) as { platform?: string };
-        platform = parsed.platform || 'web';
-      }
-    } catch (e) {
-      console.error('Error parsing OAuth state:', e);
-    }
+    const platform = parsePlatformFromState(req);
 
     const { name, emails } = profile;
